refactor(tests): update entity imports to use generated Prisma models and adjust type assertions
